up to chapter 12 rest of your life

+++ Raytracer/Raytracer_personal/include/vec3.hpp
@@ -19,16 +19,16 @@ inline double random_double(double min, double max) {
 
 class vec3 {
     public:
-        double e[3];
+        float e[3];
 
         // constructors
         vec3() : e{0, 0, 0} {}
-        vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
+        vec3(float e0, float e1, float e2) : e{e0, e1, e2} {}
 
         // position getter functions
-        double x() const { return e[0]; }
-        double y() const { return e[1]; }
-        double z() const { return e[2]; }
+        float x() const { return e[0]; }
+        float y() const { return e[1]; }
+        float z() const { return e[2]; }
 
         // utility functions
         double len() const { return std::sqrt(len_sqrd()); }
@@ -48,8 +48,8 @@ class vec3 {
 
         // operator overload functions
         vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-        double operator[](int i) const { return e[i]; }
-        double& operator[](int i) { return e[i]; }
+        float operator[](int i) const { return e[i]; }
+        float& operator[](int i) { return e[i]; }
 
         vec3& operator+=(const vec3& v) {
             e[0] += v.e[0];
